Allow to handle several UARTs/robots simultaneously
To use several UARTs, start guili like this:

  guili.py 2080 galipeur:/dev/ttyUSB0 galipette:/dev/ttyUSB1

Each UART must be associated to a robot name (here, "galipeur" and
"galipette"). If a single UART is used, name will default to "robot".

Usually, robot to be displayed/handled can be configured per portlet
using the eye icon.

In console, the "global" message name broadcasts to all robots.
To send a message to a given robot: use the same methods defined on the
object named after the robot. For instance:

  order_dummy(1, 2)         broadcast to all
  strat.order_dummy(1, 2)   send to the 'start' robot

Some portlets will need additional changes/tweaking for better
usability (e.g. per-robot color on the detection portlet).

An event is defined to update robots list, but it is not supported yet:
for now, robots are assumed to not be updated. The event is only
triggered at startup.

The following has not be tested:
- bootloader upload
- detection portlet
- use of "real" UARTs and ROME clients

Additional (minor) changes:
- frame handlers are now stored in a list (not a map).
- add an 'order_dummy' message to the test server, to test console